feat: add ignore

+++ server/api/docGenerator/pushAll.js
@@ -1,26 +1,95 @@
 import fs_sys from "fs";
 import path from "path";
-import { fs } from "@nstation/utils";
 import { promises as fs_promise } from "fs";
+import { fs, rootPath } from "@nstation/utils";
 
 const getTemplate = (table_name, fields) => {
   fields = fields.filter((field) => field?.slug !== "id");
 
-  return [
-    {
+  const docsPath = path.join(
+    rootPath,
+    "src",
+    "tables",
+    table_name,
+    "docs",
+    "index.json"
+  );
+
+  let currentTableDocs = [];
+
+  if (fs_sys.existsSync(docsPath)) {
+    currentTableDocs = fs_sys.readFileSync(docsPath, "utf-8");
+    currentTableDocs = JSON.parse(currentTableDocs);
+  }
+
+  const current_post_index = currentTableDocs.findIndex(
+    (item) => item?.method === "POST" && item?.path === `/tables/${table_name}`
+  );
+  const current_get_index = currentTableDocs.findIndex(
+    (item) => item?.method === "GET" && item?.path === `/tables/${table_name}`
+  );
+  const current_put_index = currentTableDocs.findIndex(
+    (item) =>
+      item?.method === "PUT" && item?.path === `/tables/${table_name}/:id`
+  );
+  const current_delete_index = currentTableDocs.findIndex(
+    (item) =>
+      item?.method === "DELETE" && item?.path === `/tables/${table_name}/:id`
+  );
+
+  if (current_post_index !== -1) {
+    const current_post = currentTableDocs[current_post_index];
+
+    if (!current_post?.ignore) {
+      currentTableDocs[current_post_index].body = fields?.map((field) => ({
+        name: field?.slug,
+        type: field?.type,
+        required: field?.required,
+      }));
+    }
+  } else {
+    currentTableDocs.push({
       method: "POST",
-      path: `/p/tables/${table_name}`,
-      description: "Create an entry",
+      path: `/tables/${table_name}`,
       body: fields?.map((field) => ({
         name: field?.slug,
         type: field?.type,
         required: field?.required,
       })),
-    },
-    {
+    });
+  }
+
+  if (current_get_index !== -1) {
+    const current_get = currentTableDocs[current_get_index];
+
+    if (!current_get?.ignore) {
+      currentTableDocs[current_get_index].query = [
+        {
+          name: "filters",
+          type: "string",
+          required: false,
+        },
+        {
+          name: "sort",
+          type: "string",
+          required: false,
+        },
+        {
+          name: "page",
+          type: "number",
+          required: false,
+        },
+        {
+          name: "pageSize",
+          type: "number",
+          required: false,
+        },
+      ];
+    }
+  } else {
+    currentTableDocs.push({
       method: "GET",
-      path: `/p/tables/${table_name}`,
-      description: "Get all entries",
+      path: `/tables/${table_name}`,
       query: [
         {
           name: "filters",
@@ -43,23 +112,44 @@ const getTemplate = (table_name, fields) => {
           required: false,
         },
       ],
-    },
-    {
+    });
+  }
+
+  if (current_put_index !== -1) {
+    const current_put = currentTableDocs[current_put_index];
+
+    if (!current_put?.ignore) {
+      currentTableDocs[current_put_index].body = fields?.map((field) => ({
+        name: field?.slug,
+        type: field?.type,
+        required: field?.required,
+      }));
+    }
+  } else {
+    currentTableDocs.push({
       method: "PUT",
-      path: `/p/tables/${table_name}/:entry_id`,
-      description: "Update an entry",
+      path: `/tables/${table_name}/:id`,
       body: fields?.map((field) => ({
         name: field?.slug,
         type: field?.type,
         required: field?.required,
       })),
-    },
-    {
+    });
+  }
+
+  if (current_delete_index !== -1) {
+    const current_delete = currentTableDocs[current_delete_index];
+
+    if (!current_delete?.ignore) {
+    }
+  } else {
+    currentTableDocs.push({
       method: "DELETE",
-      path: `/p/tables/${table_name}/:entry_id`,
-      description: "Delete an entry",
-    },
-  ];
+      path: `/tables/${table_name}/:id`,
+    });
+  }
+
+  return currentTableDocs;
 };
 
 const pushAll = async (table_path, regenerate = false) => {
